+++ src/Factory/ApplicationFactory.php
@@ -9,13 +9,11 @@ declare(strict_types=1);
 
 namespace Dot\Console\Factory;
 
+use Dot\Console\Application;
 use Dot\Console\Dispatcher;
-use Dot\ErrorHandler\ErrorHandler;
 use Dot\ErrorHandler\ErrorHandlerInterface;
 use Psr\Container\ContainerInterface;
 use Laminas\Console\Console;
-use Dot\Console\Application;
-use Laminas\Log\Logger;
 /**
  * Class ApplicationFactory
  * @package Dot\Console\Factory
@@ -26,22 +24,16 @@ class ApplicationFactory
      * @param ContainerInterface $container
      * @return Application
      */
-    public function __invoke(ContainerInterface $container)
+    public function __invoke(ContainerInterface $container): Application
     {
-        /** @var Logger $logger */
-        $logger = $container->get('dot-log.my_logger');
-        $dispatcher = new Dispatcher($container);
-
-        $app = new Application(
+        return new Application(
             $container->get('config')['dot_console']['name'],
             $container->get('config')['dot_console']['version'],
             $container->get('config')['dot_console']['commands'],
+            $container->get('dot-log.my_logger'),
+            $container->get(ErrorHandlerInterface::class),
             Console::getInstance(),
-            $dispatcher,
-            $logger,
-            $container->get(ErrorHandlerInterface::class)
+            new Dispatcher($container)
         );
-
-        return $app;
     }
 }